+++ ChangeLog.md
@@ -1,4 +1,7 @@
 # v2.3.0
+IMPROVEMENTS
+- allow backup and properly restore table with system.mutations is_done=0 status. fix [529](https://github.com/AlexAkulov/clickhouse-backup/issues/529)
+
 BUG FIXES
 - apply SETTINGS check_table_dependencies=0 to DROP DATABASE statement, when pass `--ignore-dependencies` together with `--rm` in `restore` command, fix [651](https://github.com/AlexAkulov/clickhouse-backup/issues/651)